Make all counts 16-bit
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@1946 42af7a65-404d-4744-a932-0658087f49c3

+++ include/nuttx/nxflat.h
@@ -81,7 +81,7 @@ struct nxflat_loadinfo_s
   /* This is temporary memory where relocation records will be loaded. */
 
   uint32 relocstart;       /* Start of array of struct flat_reloc */
-  uint32 reloccount;       /* Number of elements in reloc array */
+  uint16 reloccount;       /* Number of elements in reloc array */
 
   /* File descriptors */
